[run] Fix
Move check in make
