+++ make.py
@@ -3,6 +3,7 @@
 import sys
 sys.path.append('scripts')
 sys.path.append('scripts/develop')
+sys.path.append('scripts/develop/vendor')
 sys.path.append('scripts/core_common')
 sys.path.append('scripts/core_common/modules')
 import config
@@ -13,6 +14,7 @@ import build_server
 import deploy
 import make_common
 import config_server as develop_config_server
+import dependence
 
 # parse configuration
 config.parse()
@@ -79,6 +81,8 @@ base.configure_common_apps()
 
 # developing...
 if ("1" == config.option("develop")):
+  if not dependence.check_dependencies():
+    sys.exit()
   build_server.build_server_develop()
   build_js.build_js_develop(base_dir + "/..")
   develop_config_server.make()
