+++ scripts/develop/run.py
@@ -1,21 +1,10 @@
 import sys
 sys.path.append('../')
-sys.path.append('vendor')
 import os
 import base
 import dependence
 import traceback
 import config
-
-platform = base.host_platform()
-
-if ("windows" == platform):
-  import libwindows
-
-platform = base.host_platform()
-
-if ("windows" == platform):
-  import libwindows
 
 if (sys.version_info[0] >= 3):
   unicode = str
@@ -50,57 +39,20 @@ def start_mac_services():
 def run_integration_example():
   base.cmd_in_dir('../../../document-server-integration/web/documentserver-example/nodejs', 'python', ['run-develop.py'])
 
-def check_dependencies():
-  checksResult = dependence.CDependencies()
-  
-  checksResult.append(dependence.check_git())
-  checksResult.append(dependence.check_nodejs())
-  checksResult.append(dependence.check_java())
-  checksResult.append(dependence.check_erlang())
-  checksResult.append(dependence.check_rabbitmq())
-  checksResult.append(dependence.check_gruntcli())
-  checksResult.append(dependence.check_buildTools())
-  checksResult.append(dependence.check_mysqlServer())
-  
-  server_addons = []
-  if (config.option("server-addons") != ""):
-    server_addons = config.option("server-addons").rsplit(", ")
-  if ("server-lockstorage" in server_addons):
-    checksResult.append(dependence.check_redis())
-  
-  if (len(checksResult.install) > 0):
-    install_args = ['install.py']
-    install_args += checksResult.get_uninstall()
-    install_args += checksResult.get_removepath()
-    install_args += checksResult.get_install()
-    install_args += ['--mysql-path', unicode(checksResult.mysqlPath)]
-    code = libwindows.sudo(unicode(sys.executable), install_args)
-  
-  return dependence.check_MySQLConfig(checksResult.mysqlPath)
-
 def make(args = []):
   try:
-    base.cmd_in_dir('../../', 'python', ['configure.py', '--branch', 'develop', '--develop', '1', '--module', 'server', '--update', '1', '--update-light', '1', '--clean', '0'] + args)
-    
-    # parse configuration
-    config.parse()
-    # correct defaults (the branding repo is already updated)
-    config.parse_defaults()
-    
     base.configure_common_apps()
-  
+    
+    platform = base.host_platform()
     if ("windows" == platform):
       dependence.check_pythonPath()
-      if not dependence.check_vc_components():
-        sys.exit()
-      if not check_dependencies():
-        sys.exit()
       restart_win_rabbit()
     elif ("mac" == platform):
       start_mac_services()
     
     dependence.check_gitPath()
     base.print_info('Build modules')
+    base.cmd_in_dir('../../', 'python', ['configure.py', '--branch', 'develop', '--develop', '1', '--module', 'server', '--update', '1', '--update-light', '1', '--clean', '0'] + args)
     base.cmd_in_dir('../../', 'python', ['make.py'])
   
     run_integration_example()
